+++ src/npc/plugin.rs
@@ -4,7 +4,7 @@ use bevy::prelude::*;
 use crate::{
     npc::{
         components::{NpcIdGenerator, ScheduleTicker},
-        systems::{spawn_debug_npcs, tick_schedule_state},
+        systems::{drive_npc_locomotion, spawn_debug_npcs, tick_schedule_state},
     },
     world::systems::spawn_world_environment,
 };
@@ -16,6 +16,6 @@ impl Plugin for NpcPlugin {
         app.init_resource::<NpcIdGenerator>()
             .init_resource::<ScheduleTicker>()
             .add_systems(Startup, spawn_debug_npcs.after(spawn_world_environment))
-            .add_systems(Update, tick_schedule_state);
+            .add_systems(Update, (tick_schedule_state, drive_npc_locomotion));
     }
 }